test(chip): migrate basic
