+++ core/src/components/chip/test/basic/chip.e2e.ts-snapshots/chip-basic-md-ltr-Mobile-Firefox-linux.png
