+++ core/src/components/chip/test/basic/chip.e2e.ts-snapshots/chip-basic-md-rtl-Mobile-Chrome-linux.png
